Removed the helper procedures and defined the calculations which they had performed as discrete definitions.
This should correct the errors mit-scheme was giving about line not being defined.  (Because mit-scheme gave no location information on this error, the changes committed here are my best guess as to the cause of the error.  It seems to me that defining skew inside of a let statement was causing some scope issues.)